Move command to runner package
