+++ pkg/runner/command.go
@@ -1,4 +1,4 @@
-package command
+package runner
 
 import (
 	"context"
