+++ pkg/runner/runner.go
@@ -1,8 +1,6 @@
 package runner
 
-import "github.com/supergiant/supergiant/pkg/runner/command"
-
 // Runner is interface for running command in different environment
 type Runner interface {
-	Run(command *command.Command) error
+	Run(command *Command) error
 }
